Adding support for Azure openAI
