+++ Sources/OpenAI/Public/Service/OpenAIServiceFactory.swift
@@ -31,4 +31,24 @@ public class OpenAIServiceFactory {
          configuration: configuration,
          decoder: decoder)
    }
+   
+   /// Creates and returns an instance of `OpenAIService`.
+   ///
+   /// - Parameters:
+   ///   - azureConfiguration: The AzureOpenAIConfiguration.
+   ///   - urlSessionConfiguration: The URL session configuration to be used for network calls (default is `.default`).
+   ///   - decoder: The JSON decoder to be used for parsing API responses (default is `JSONDecoder.init()`).
+   ///
+   /// - Returns: A fully configured object conforming to `OpenAIService`.
+   public static func service(
+      azureConfiguration: AzureOpenAIConfiguration,
+      urlSessionConfiguration: URLSessionConfiguration = .default,
+      decoder: JSONDecoder = .init())
+   -> some OpenAIService
+   {
+      DefaultOpenAIAzureService(
+         azureConfiguration: azureConfiguration,
+         urlSessionConfiguration: urlSessionConfiguration,
+         decoder: decoder)
+   }
 }
